+++ context_pipe/__init__.py
@@ -2,5 +2,5 @@
 # Copyright (c) 2026 Luis Kobayashi. All rights reserved.
 from .api import pipe
 
-__version__ = "0.4.7"
+__version__ = "0.5.0"
 __all__ = ["pipe", "__version__"]